Refactor UpdateReviewDto and RoomController to handle optional images

+++ src/room/room.service.ts
@@ -602,7 +602,7 @@ export class RoomService {
   async updateReview(
     id: number,
     data: UpdateReviewDto,
-    images: Express.Multer.File[],
+    images?: Express.Multer.File[],
   ): Promise<Review> {
     const oldReview = await this.prisma.review.findUnique({
       where: { id },
@@ -618,7 +618,7 @@ export class RoomService {
       });
     }
     const { content, star } = data;
-    const review_image: ReviewImage[] = [];
+    let review_image: ReviewImage[] = [];
     const uploadedUrls: string[] = [];
     try {
       const newReview = await this.prisma.$transaction(
@@ -627,29 +627,34 @@ export class RoomService {
             where: { id },
             data: { content, star },
           });
-          for (const image of images) {
-            const key = `review/${review.room_id}/${getKeyByFilename(
-              image.originalname,
-            )}`;
-            const { url } = await this.uploadService.uploadFile(image, key);
-            this.logger.log(`Uploaded ${url}`);
-            uploadedUrls.push(url);
-            const newReviewImage = await prisma.reviewImage.create({
-              data: {
-                review_id: review.id,
-                image_url: url,
-              },
-            });
-            review_image.push(newReviewImage);
+          if (images && images.length > 0) {
+            for (const image of images) {
+              const key = `review/${review.room_id}/${getKeyByFilename(
+                image.originalname,
+              )}`;
+              const { url } = await this.uploadService.uploadFile(image, key);
+              this.logger.log(`Uploaded ${url}`);
+              uploadedUrls.push(url);
+              const newReviewImage = await prisma.reviewImage.create({
+                data: {
+                  review_id: review.id,
+                  image_url: url,
+                },
+              });
+              review_image.push(newReviewImage);
+            }
+
+            for (const image of oldReview.review_image) {
+              await this.uploadService.deleteFileS3(image.image_url);
+              this.logger.log(`Deleted ${image.image_url}`);
+              await this.prisma.reviewImage.delete({
+                where: { id: image.id },
+              });
+            }
+          } else {
+            review_image = oldReview.review_image;
           }
 
-          for (const image of oldReview.review_image) {
-            await this.uploadService.deleteFileS3(image.image_url);
-            this.logger.log(`Deleted ${image.image_url}`);
-            await this.prisma.reviewImage.delete({
-              where: { id: image.id },
-            });
-          }
           return {
             ...review,
             review_image,
